Separate passport-provider specific code
- Move passport-related methods from ProviderService
  to PassportProviderService
- Make ProviderService an abstract class
- Generalize createLinkInternal
- Move createLink into PassportProviderService
  (token and fence account services will have different
   implementations)
- Move audit logging out of controller createLink
