+++ service/src/main/java/bio/terra/externalcreds/services/ProviderService.java
@@ -12,9 +12,7 @@ import bio.terra.externalcreds.models.CannotDecodeOAuth2State;
 import bio.terra.externalcreds.models.LinkedAccount;
 import bio.terra.externalcreds.models.LinkedAccountWithPassportAndVisas;
 import bio.terra.externalcreds.models.OAuth2State;
-import bio.terra.externalcreds.models.VisaVerificationDetails;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.annotations.VisibleForTesting;
 import java.security.SecureRandom;
 import java.sql.Timestamp;
 import java.time.Duration;
@@ -24,34 +22,30 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
 import org.springframework.security.oauth2.core.OAuth2AuthorizationException;
 import org.springframework.security.oauth2.core.OAuth2ErrorCodes;
-import org.springframework.security.oauth2.core.OAuth2RefreshToken;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
 @Service
 @Slf4j
-public class ProviderService {
-  private final ExternalCredsConfig externalCredsConfig;
-  private final ProviderClientCache providerClientCache;
-  private final OAuth2Service oAuth2Service;
-  private final LinkedAccountService linkedAccountService;
-  private final PassportService passportService;
-  private final JwtUtils jwtUtils;
-  private final AuditLogger auditLogger;
-  private final SecureRandom secureRandom = new SecureRandom();
-  private final ObjectMapper objectMapper;
-  private static final Collection<String> unrecoverableOAuth2ErrorCodes =
+public abstract class ProviderService {
+  public final ExternalCredsConfig externalCredsConfig;
+  public final ProviderClientCache providerClientCache;
+  public final OAuth2Service oAuth2Service;
+  public final LinkedAccountService linkedAccountService;
+  public final AuditLogger auditLogger;
+  public final SecureRandom secureRandom = new SecureRandom();
+  public final ObjectMapper objectMapper;
+  public static final Collection<String> unrecoverableOAuth2ErrorCodes =
       Set.of(
           OAuth2ErrorCodes.ACCESS_DENIED,
           OAuth2ErrorCodes.INSUFFICIENT_SCOPE,
@@ -71,44 +65,18 @@ public class ProviderService {
       ProviderClientCache providerClientCache,
       OAuth2Service oAuth2Service,
       LinkedAccountService linkedAccountService,
-      PassportService passportService,
-      JwtUtils jwtUtils,
       AuditLogger auditLogger,
       ObjectMapper objectMapper) {
     this.externalCredsConfig = externalCredsConfig;
     this.providerClientCache = providerClientCache;
     this.oAuth2Service = oAuth2Service;
     this.linkedAccountService = linkedAccountService;
-    this.passportService = passportService;
-    this.jwtUtils = jwtUtils;
     this.auditLogger = auditLogger;
     this.objectMapper = objectMapper;
   }
 
   public Set<String> getProviderList() {
     return Collections.unmodifiableSet(externalCredsConfig.getProviders().keySet());
-  }
-
-  /**
-   * Get a new passport for each linked accounts with visas or passports expiring within
-   * externalCredsConfig.getVisaAndPassportRefreshInterval time from now
-   *
-   * @return the number of linked accounts with expiring visas or passports
-   */
-  public int refreshExpiringPassports() {
-    var refreshInterval = externalCredsConfig.getVisaAndPassportRefreshDuration();
-    var expirationCutoff = new Timestamp(Instant.now().plus(refreshInterval).toEpochMilli());
-    var expiringLinkedAccounts = linkedAccountService.getExpiringLinkedAccounts(expirationCutoff);
-
-    for (LinkedAccount linkedAccount : expiringLinkedAccounts) {
-      try {
-        authAndRefreshPassport(linkedAccount);
-      } catch (Exception e) {
-        log.info("Failed to refresh passport, will try again at the next interval.", e);
-      }
-    }
-
-    return expiringLinkedAccounts.size();
   }
 
   public Optional<String> getProviderAuthorizationUrl(
@@ -148,32 +116,7 @@ public class ProviderService {
     }
   }
 
-  public Optional<LinkedAccountWithPassportAndVisas> createLink(
-      String providerName, String userId, String authorizationCode, String encodedState) {
-
-    var oAuth2State = validateOAuth2State(providerName, userId, encodedState);
-
-    return providerClientCache
-        .getProviderClient(providerName)
-        .map(
-            providerClient -> {
-              var providerInfo = externalCredsConfig.getProviders().get(providerName);
-              try {
-                return createLinkInternal(
-                    providerName,
-                    userId,
-                    authorizationCode,
-                    oAuth2State.getRedirectUri(),
-                    new HashSet<>(providerInfo.getScopes()),
-                    encodedState,
-                    providerClient);
-              } catch (OAuth2AuthorizationException oauthEx) {
-                throw new BadRequestException(oauthEx);
-              }
-            });
-  }
-
-  private OAuth2State validateOAuth2State(String providerName, String userId, String encodedState) {
+  public OAuth2State validateOAuth2State(String providerName, String userId, String encodedState) {
     try {
       OAuth2State oAuth2State = OAuth2State.decode(objectMapper, encodedState);
       if (!providerName.equals(oAuth2State.getProvider())) {
@@ -186,7 +129,7 @@ public class ProviderService {
     }
   }
 
-  private LinkedAccountWithPassportAndVisas createLinkInternal(
+  public ImmutablePair<LinkedAccount, OAuth2User> createLinkInternal(
       String providerName,
       String userId,
       String authorizationCode,
@@ -222,8 +165,7 @@ public class ProviderService {
               "user info from provider %s did not contain external id claim %s",
               providerName, providerInfo.getExternalIdClaim()));
     }
-    var linkedAccount =
-        new LinkedAccount.Builder()
+    LinkedAccount linkedAccount = new LinkedAccount.Builder()
             .providerName(providerName)
             .userId(userId)
             .expires(expires)
@@ -231,10 +173,32 @@ public class ProviderService {
             .refreshToken(refreshToken.getTokenValue())
             .isAuthenticated(true)
             .build();
-
-    return linkedAccountService.upsertLinkedAccountWithPassportAndVisas(
-        jwtUtils.enrichAccountWithPassportAndVisas(linkedAccount, userInfo));
+    return new ImmutablePair<>(linkedAccount, userInfo);
   }
+
+  /**
+   * Traverse causes for oauthEx until the cause is null or there is a cycle in the causes. Return
+   * the last oauth2 error code found. This is needed because Spring likes to wrap
+   * OAuth2AuthorizationExceptions with other OAuth2AuthorizationExceptions that have non-standard
+   * error codes. We just want to handle standard error codes which should be in the root cause.
+   */
+  public String getRootOAuth2ErrorCode(OAuth2AuthorizationException oauthEx) {
+    var errorCode = oauthEx.getError().getErrorCode();
+    Throwable currentThrowable = oauthEx.getCause();
+    var visitedThrowables = new ArrayList<Throwable>();
+    visitedThrowables.add(oauthEx);
+
+    while (currentThrowable != null && !visitedThrowables.contains(currentThrowable)) {
+      if (currentThrowable instanceof OAuth2AuthorizationException nestedOauthEx) {
+        errorCode = nestedOauthEx.getError().getErrorCode();
+      }
+      visitedThrowables.add(currentThrowable);
+      currentThrowable = currentThrowable.getCause();
+    }
+
+    return errorCode;
+  }
+
 
   public LinkedAccount deleteLink(String userId, String providerName) {
     var providerInfo = externalCredsConfig.getProviders().get(providerName);
@@ -255,178 +219,8 @@ public class ProviderService {
     return linkedAccount;
   }
 
-  public int validateAccessTokenVisas() {
-    var visaDetailsList = passportService.getUnvalidatedAccessTokenVisaDetails();
 
-    var linkedAccountIdsToRefresh =
-        visaDetailsList.stream()
-            .flatMap(
-                visaDetails ->
-                    validateVisaWithProvider(visaDetails)
-                        ? Stream.empty()
-                        : Stream.of(visaDetails.getLinkedAccountId()))
-            .distinct();
-
-    linkedAccountIdsToRefresh.forEach(
-        linkedAccountId -> {
-          var linkedAccount = linkedAccountService.getLinkedAccount(linkedAccountId);
-          try {
-            linkedAccount.ifPresentOrElse(
-                this::authAndRefreshPassport,
-                () -> log.info("No linked account found when trying to validate passport."));
-          } catch (Exception e) {
-            log.info("Failed to refresh passport, will try again at the next interval.", e);
-          }
-        });
-
-    return visaDetailsList.size();
-  }
-
-  @VisibleForTesting
-  void authAndRefreshPassport(LinkedAccount linkedAccount) {
-    if (linkedAccount.getExpires().toInstant().isBefore(Instant.now())) {
-      invalidateLinkedAccount(linkedAccount);
-    } else {
-      try {
-        var linkedAccountWithRefreshedPassport = getRefreshedPassportsAndVisas(linkedAccount);
-        linkedAccountService.upsertLinkedAccountWithPassportAndVisas(
-            linkedAccountWithRefreshedPassport);
-        var transactionClaim =
-            linkedAccountWithRefreshedPassport
-                .getPassport()
-                .flatMap(p -> jwtUtils.getJwtTransactionClaim(p.getJwt()));
-        auditLogger.logEvent(
-            new AuditLogEvent.Builder()
-                .auditLogEventType(AuditLogEventType.LinkRefreshed)
-                .providerName(linkedAccount.getProviderName())
-                .userId(linkedAccount.getUserId())
-                .externalUserId(linkedAccount.getExternalUserId())
-                .transactionClaim(transactionClaim)
-                .build());
-
-      } catch (IllegalArgumentException iae) {
-        throw new ExternalCredsException(
-            String.format(
-                "Could not contact issuer for provider %s", linkedAccount.getProviderName()),
-            iae);
-      } catch (OAuth2AuthorizationException oauthEx) {
-        // if it looks like the refresh token will never work, delete the passport
-        if (unrecoverableOAuth2ErrorCodes.contains(getRootOAuth2ErrorCode(oauthEx))) {
-          log.info(
-              String.format(
-                  "Caught unrecoverable oauth2 error code refreshing passport for user id [%s].",
-                  linkedAccount.getUserId()),
-              oauthEx);
-          if (linkedAccount.getId().isEmpty()) {
-            throw new ExternalCredsException("linked account id missing");
-          }
-          invalidateLinkedAccount(linkedAccount);
-        } else {
-          // log and try again later
-          throw new ExternalCredsException("Failed to refresh passport: ", oauthEx);
-        }
-      }
-    }
-  }
-
-  /**
-   * Traverse causes for oauthEx until the cause is null or there is a cycle in the causes. Return
-   * the last oauth2 error code found. This is needed because Spring likes to wrap
-   * OAuth2AuthorizationExceptions with other OAuth2AuthorizationExceptions that have non-standard
-   * error codes. We just want to handle standard error codes which should be in the root cause.
-   */
-  private String getRootOAuth2ErrorCode(OAuth2AuthorizationException oauthEx) {
-    var errorCode = oauthEx.getError().getErrorCode();
-    Throwable currentThrowable = oauthEx.getCause();
-    var visitedThrowables = new ArrayList<Throwable>();
-    visitedThrowables.add(oauthEx);
-
-    while (currentThrowable != null && !visitedThrowables.contains(currentThrowable)) {
-      if (currentThrowable instanceof OAuth2AuthorizationException nestedOauthEx) {
-        errorCode = nestedOauthEx.getError().getErrorCode();
-      }
-      visitedThrowables.add(currentThrowable);
-      currentThrowable = currentThrowable.getCause();
-    }
-
-    return errorCode;
-  }
-
-  private LinkedAccountWithPassportAndVisas getRefreshedPassportsAndVisas(
-      LinkedAccount linkedAccount) {
-    var clientRegistration =
-        providerClientCache
-            .getProviderClient(linkedAccount.getProviderName())
-            .orElseThrow(
-                () ->
-                    new ExternalCredsException(
-                        String.format(
-                            "Unable to find configs for the provider: %s",
-                            linkedAccount.getProviderName())));
-    var accessTokenResponse =
-        oAuth2Service.authorizeWithRefreshToken(
-            clientRegistration, new OAuth2RefreshToken(linkedAccount.getRefreshToken(), null));
-
-    // save the linked account with the new refresh token and extracted passport
-    var linkedAccountWithRefreshToken =
-        Optional.ofNullable(accessTokenResponse.getRefreshToken())
-            .map(
-                refreshToken ->
-                    linkedAccountService.upsertLinkedAccount(
-                        linkedAccount.withRefreshToken(refreshToken.getTokenValue())))
-            .orElse(linkedAccount);
-
-    // update the passport and visas
-    var userInfo =
-        oAuth2Service.getUserInfo(clientRegistration, accessTokenResponse.getAccessToken());
-    return jwtUtils.enrichAccountWithPassportAndVisas(linkedAccountWithRefreshToken, userInfo);
-  }
-
-  @VisibleForTesting
-  boolean validateVisaWithProvider(VisaVerificationDetails visaDetails) {
-    var providerProperties = externalCredsConfig.getProviders().get(visaDetails.getProviderName());
-    if (providerProperties == null) {
-      throw new NotFoundException(
-          String.format("Provider %s not found", visaDetails.getProviderName()));
-    }
-
-    var validationEndpoint =
-        providerProperties
-            .getValidationEndpoint()
-            .orElseThrow(
-                () ->
-                    new NotFoundException(
-                        String.format(
-                            "Validation endpoint for provider %s not found",
-                            visaDetails.getProviderName())));
-
-    var response =
-        WebClient.create(validationEndpoint)
-            .get()
-            .uri(uriBuilder -> uriBuilder.queryParam("visa", visaDetails.getVisaJwt()).build())
-            .retrieve();
-    var responseBody =
-        response
-            .onStatus(HttpStatusCode::isError, clientResponse -> Mono.empty())
-            .bodyToMono(String.class)
-            .block(Duration.of(1000, ChronoUnit.MILLIS));
-
-    log.info(
-        "Got visa validation response.",
-        Map.of(
-            "linkedAccountId", visaDetails.getLinkedAccountId(),
-            "providerName", visaDetails.getProviderName(),
-            "validationResponse", Objects.requireNonNullElse(responseBody, "[null]")));
-
-    var visaValid = "valid".equalsIgnoreCase(responseBody);
-
-    if (visaValid) {
-      passportService.updateVisaLastValidated(visaDetails.getVisaId());
-    }
-    return visaValid;
-  }
-
-  private void invalidateLinkedAccount(LinkedAccount linkedAccount) {
+  public void invalidateLinkedAccount(LinkedAccount linkedAccount) {
     auditLogger.logEvent(
         new AuditLogEvent.Builder()
             .auditLogEventType(AuditLogEventType.LinkExpired)
